(PC-15447)[PRO] chore: extract deactivation dialog out of offers ActionsBar
