+++ pro/src/components/pages/Offers/Offers/ActionsBar/ConfirmDialog.tsx/DeactivationConfirmDialog/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DeactivationConfirmDialog'
